// FrontMiJardin/src/app/interfaces/DatosNotaContableReportes.ts
export class DatosNotaContableReportes {
        idNotaContable: number ;
        usuarioCreacion: string ;
        nombreCliente: string ;
        razon: string ;
        unidadComercial: string ;
        pais: string ;
        entidadLegal: string ;
        tipoMoneda: string ;
        negocio: string ;
        concepto: string ;
        region: string ;
        udn: string ;
        estado: string ;
        aprobador: string ;
        tipoNota: string ;
        usd: number ;
        producto: string ;
        descripcion: string ;
        fechaCreacion: string ;
        fechaAprobacion: string ;
        observacionAprobacion: string;
        listaAprobadores: Array<string>;
    }